packages/clawteam/pyproject.toml
[project]
name = "clawteam"
version = "0.1.1"
description = "Framework-agnostic multi-agent coordination CLI"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "HKUDS", email = "jiabintang77@gmail.com"}
]
keywords = ["ai", "agent", "multi-agent", "team", "coordination", "cli", "swarm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "typer>=0.12.0,<1.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "rich>=13.0.0,<15.0.0",
    "tomli>=2.0.0; python_version < '3.11'",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0,<10.0.0",
    "ruff>=0.1.0",
]
p2p = [
    "pyzmq>=25.0.0,<27.0.0",
]

[project.scripts]
clawteam = "clawteam.cli.commands:app"

[project.urls]
Homepage = "https://github.com/HKUDS/ClawTeam"
Repository = "https://github.com/HKUDS/ClawTeam"
Issues = "https://github.com/HKUDS/ClawTeam/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["clawteam"]

[tool.hatch.build.targets.sdist]
include = [
    "clawteam/",
    "README.md",
    "LICENSE",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
